% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modern_to_plot.R
\name{add_modern_to_plot}
\alias{add_modern_to_plot}
\title{Adds modern data to ice core plot}
\usage{
add_modern_to_plot(ice_core, modern_polar, modern_global)
}
\description{
Adds modern temp and co2 data to ice core plot (Penny Rowe - Polar project)
}
