% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom}
\alias{binom}
\title{Binomial Distribution Calculations}
\usage{
binom()
}
\description{
This function allows to compute quantities relating to binomial distributions and visualize the results clearly.
}
