% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_t_test.R
\name{conduct_t_test}
\alias{conduct_t_test}
\title{Conduct a t-test using provided test statistic data.}
\usage{
conduct_t_test(t, df, type)
}
\arguments{
\item{t}{the t-statistic you are testing}

\item{df}{the degrees of freedom in your sample}

\item{type}{the type of test. Options are 'less', 'greater', and 'both'}
}
\description{
This function conducts a t-test, exports the results, exports what a user would need to type into R to get the same result, and produces a visualization.
}
\examples{
> conduct_t_test(3,10,"greater") 
$prob
[1] 0.006671828

$p_value_type
[1] "1-pt(3,10)"
}
