% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_value_activity.R
\name{expected_value_activity}
\alias{expected_value_activity}
\title{Flip a coin three times for every student}
\usage{
expected_value_activity()
}
\description{
Gives everyone in class a number of heads flipped out of three (in case they don't have coins for the expected value activity).
}
