% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ice_core_T_and_co2_on_same_grid.R
\name{get_ice_core_T_and_co2_on_same_grid}
\alias{get_ice_core_T_and_co2_on_same_grid}
\title{Put Ice Core Data on same grid}
\usage{
get_ice_core_T_and_co2_on_same_grid(ice_core_co2, ice_core_T)
}
\description{
Puts modern ice core temp and co2 data on the same time grid (Penny Rowe - Polar project)
}
