% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modern_global_T_anomaly_and_co2.R
\name{get_modern_global_T_anomaly_and_co2}
\alias{get_modern_global_T_anomaly_and_co2}
\title{Modern Anomaly Data}
\usage{
get_modern_global_T_anomaly_and_co2(modern_co2m, law_dome_co2)
}
\description{
Loads modern temperature anomaly and CO2 data (Penny Rowe - Polar Project)
}
