% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modern_polar_T_and_co2_on_same_grid.R
\name{get_modern_polar_T_and_co2_on_same_grid}
\alias{get_modern_polar_T_and_co2_on_same_grid}
\title{Get Same Grid for Polar Data}
\usage{
get_modern_polar_T_and_co2_on_same_grid(modern_polar_T, modern_co2)
}
\description{
Places polar temperature and CO2 data on same time grid.
}
