% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalProblem.R
\name{hospitalProblem}
\alias{hospitalProblem}
\title{Creates tables for the hospital contingency table problem}
\usage{
hospitalProblem()
}
\description{
Creates tables for the hospital contingency table problem
}
