% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_test_activity.R
\name{hyp_test_activity}
\alias{hyp_test_activity}
\title{Guide the class in conducting hypothesis tests}
\usage{
hyp_test_activity(n = 9, numstudents = "N")
}
\description{
Allows students to calculate their own sample mean and use it to do a significance test.
}
