% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariaLoop.R
\name{malariaLoop}
\alias{malariaLoop}
\title{Run a randomization simulation of the malaria activity in the text}
\usage{
malariaLoop(reps = 10000)
}
\description{
This function runs a randomization simulation to show the 
likelihood of different scenarios under the null hypothesis
for the malaria example from the text. It produces a bar plot
and an assessment of the frequency of the observed result.
}
\examples{
> malariaLoop(reps = 10000)
}
