% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_ice_core_co2_and_T.R
\name{plot_timeseries_ice_core_co2_and_T}
\alias{plot_timeseries_ice_core_co2_and_T}
\title{Ice Core Timeseries}
\usage{
plot_timeseries_ice_core_co2_and_T(ice_core_co2, ice_core_T)
}
\description{
Plots simultaneous time series for ice core data.
}
