% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_modern_co2_and_T.R
\name{plot_timeseries_modern_co2_and_T}
\alias{plot_timeseries_modern_co2_and_T}
\title{Plot Modern Timeseries}
\usage{
plot_timeseries_modern_co2_and_T(df)
}
\description{
Plots simultaneous timeseries of CO2 and temperature (Penny Rowe - Polar project).
}
