% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove Outliers}
\usage{
remove_outliers()
}
\description{
This function allows you to remove the outliers found in a list you are working with and save it to a new list.
}
\examples{
> x = c(-5,1,3,4,5,5,5,7,8,8,9,10,10,15,24,27)
> new_x = remove_outliers()
Remember! remove_outliers() needs to know what you want the new list name to be!
Make sure you typed *your new list name here* = remove_outliers(). If not, press Esc to cancel.
  
What is the name of the list with your data? x

Look at the boxplot that was just created.
What is the lower cutoff for small outliers? You may answer 'none'. 0
What is the upper cutoff for large outliers? You may answer 'none'. 20


You can do this yourself by typing:
*new list name* = x[x > 0 & x < 20]
> new_x
[1]  1  3  4  5  5  5  7  8  8  9 10 10 15
}
