% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_dist_activity.R
\name{sampling_dist_activity}
\alias{sampling_dist_activity}
\title{Give everyone a sample of n heights}
\usage{
sampling_dist_activity(n = 3, numstudents = "N")
}
\description{
Allows students to calculate their own sample mean and see the sampling distribution in action.
}
