% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_dist_demo.R
\name{t_dist_demo}
\alias{t_dist_demo}
\title{Demo of Student's t-distribution}
\usage{
t_dist_demo(n = 4, numstudents = "N", section_name = "section")
}
\description{
This function allows you to produce samples for every student in a class as a way of seeing
the shape of Student's t-distribution.
}
