% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_install.R
\name{test_install}
\alias{test_install}
\title{Test That Your Installation Was Successful}
\usage{
test_install()
}
\description{
Prints a welcome banner to indicate that the installation proceeded successfully
}
\examples{
> test_install()
}
