% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_code.R
\name{update_code}
\alias{update_code}
\title{Update Code to Latest Version}
\usage{
update_code()
}
\description{
This function updates your library to the newest version and sets those to become available.
}
\examples{
> update_code()
}
