% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_dist_demo.R
\name{t_dist_demo}
\alias{t_dist_demo}
\title{T Test Demo}
\usage{
t_dist_demo(n = 6, samples = 10000)
}
\description{
This function allows me to run a demo to illustrate the need for a non-normal t-distribution for small samples.
}
